Make the Cancel button take you back home
This commit basically undoes an earlier one,
in which I had the Cancel button on the comment
form take you back to the previous page (post).

The issue occurs when the user clicking comment
button had not been logged in, and then did log
in. Hitting Cancel would take them back to the
previous page, which in this case is the login.
And that's not good, because they already logged
in once.

Now it takes you back to the home page, like it
initially used to do.