Turn sidebar_posts into a global variable
Almost every template needs the sidebar posts.
The approach until now was to pass it as an arg
to every route whose template needs it.
Which is repetitive and ugly.
I used Flask.context_processor to make the value
available to all the templates. So now,
it doesn't have to be passed to the routes.